refactor: NetworkModule 적용 및 사용자 정보 관리 구조 개선
