+++ android/Staccato_AN/app/src/main/java/com/on/staccato/data/login/LoginDefaultRepository.kt
@@ -1,6 +1,6 @@
 package com.on.staccato.data.login
 
-import com.on.staccato.StaccatoApplication
+import com.on.staccato.data.member.MemberDataSource
 import com.on.staccato.data.network.ApiResult
 import com.on.staccato.data.network.handle
 import com.on.staccato.domain.repository.LoginRepository
@@ -9,15 +9,16 @@ import javax.inject.Inject
 class LoginDefaultRepository
     @Inject
     constructor(
+        private val memberDataSource: MemberDataSource,
         private val loginDataSource: LoginDataSource,
     ) : LoginRepository {
         override suspend fun loginWithNickname(nickname: String): ApiResult<Unit> =
             loginDataSource.requestLoginWithNickname(nickname).handle {
-                StaccatoApplication.userInfoPrefsManager.setTokenAndId(it.token, it.id)
+                memberDataSource.setTokenAndId(it.token, it.id)
             }
 
         override suspend fun getToken(): Result<String?> =
             runCatching {
-                StaccatoApplication.userInfoPrefsManager.getToken()
+                memberDataSource.getToken()
             }
     }
