+++ android/Staccato_AN/app/src/main/java/com/on/staccato/data/member/MemberDataSource.kt
@@ -1,8 +1,17 @@
-package com.on.staccato.data.mypage
+package com.on.staccato.data.member
 
 import com.on.staccato.domain.model.MemberProfile
 
-interface MyPageLocalDataSource {
+interface MemberDataSource {
+    suspend fun getToken(): String?
+
+    suspend fun getMemberId(): Long
+
+    suspend fun setTokenAndId(
+        newToken: String,
+        id: Long,
+    )
+
     suspend fun getMemberProfile(): MemberProfile
 
     suspend fun updateMemberProfile(memberProfile: MemberProfile)
