+++ android/Staccato_AN/app/src/main/java/com/on/staccato/data/mypage/MyPageDataSource.kt
@@ -1,0 +1,12 @@
+package com.on.staccato.data.mypage
+
+import com.on.staccato.data.dto.mypage.MemberProfileResponse
+import com.on.staccato.data.dto.mypage.ProfileImageResponse
+import com.on.staccato.data.network.ApiResult
+import okhttp3.MultipartBody
+
+interface MyPageDataSource {
+    suspend fun loadMemberProfile(): ApiResult<MemberProfileResponse>
+
+    suspend fun updateProfileImage(profileImageFile: MultipartBody.Part): ApiResult<ProfileImageResponse>
+}
