+++ android/Staccato_AN/app/src/main/java/com/on/staccato/data/network/ApiResultCall.kt
@@ -1,20 +1,15 @@
 package com.on.staccato.data.network
 
-import com.on.staccato.StaccatoApplication.Companion.retrofit
-import com.on.staccato.data.dto.ErrorResponse
-import com.on.staccato.data.dto.Status
-import com.on.staccato.data.network.StaccatoClient.getErrorResponse
 import okhttp3.Request
-import okhttp3.ResponseBody
 import okio.Timeout
 import retrofit2.Call
-import retrofit2.HttpException
 import retrofit2.Response
-import java.io.IOException
 
 class ApiResultCall<T : Any>(
     private val delegate: Call<T>,
 ) : Call<ApiResult<T>> {
+    private val responseHandler = ResponseHandler()
+
     override fun enqueue(callback: retrofit2.Callback<ApiResult<T>>) {
         delegate.enqueue(
             object : retrofit2.Callback<T> {
@@ -22,7 +17,7 @@ class ApiResultCall<T : Any>(
                     call: Call<T>,
                     response: Response<T>,
                 ) {
-                    val networkResult: ApiResult<T> = handleApiResponse { response }
+                    val networkResult: ApiResult<T> = responseHandler.handleApiResponse { response }
                     callback.onResponse(this@ApiResultCall, Response.success(networkResult))
                 }
 
@@ -30,7 +25,7 @@ class ApiResultCall<T : Any>(
                     call: Call<T>,
                     throwable: Throwable,
                 ) {
-                    val exception = handleException<T>(throwable)
+                    val exception = responseHandler.handleException<T>(throwable)
                     callback.onResponse(this@ApiResultCall, Response.success(exception))
                 }
             },
@@ -53,38 +48,3 @@ class ApiResultCall<T : Any>(
 
     override fun timeout(): Timeout = delegate.timeout()
 }
-
-private const val CREATED = 201
-private const val NOT_FOUND_ERROR_BODY = "errorBody를 찾을 수 없습니다."
-
-private fun <T : Any> handleApiResponse(execute: () -> Response<T>): ApiResult<T> {
-    return try {
-        val response: Response<T> = execute()
-        val body: T? = response.body()
-
-        when {
-            response.isSuccessful && response.code() == CREATED -> Success(body as T)
-            response.isSuccessful && body != null -> Success(body)
-            else -> {
-                val errorBody: ResponseBody =
-                    response.errorBody()
-                        ?: throw IllegalArgumentException(NOT_FOUND_ERROR_BODY)
-                val errorResponse: ErrorResponse = retrofit.getErrorResponse(errorBody)
-                ServerError(
-                    status = Status.Message(errorResponse.status),
-                    message = errorResponse.message,
-                )
-            }
-        }
-    } catch (httpException: HttpException) {
-        ServerError(status = Status.Code(httpException.code()), message = httpException.message())
-    } catch (throwable: Throwable) {
-        handleException<T>(throwable)
-    }
-}
-
-private fun <T : Any> handleException(throwable: Throwable) =
-    when (throwable) {
-        is IOException -> Exception.NetworkError<T>()
-        else -> Exception.UnknownError<T>()
-    }
