+++ android/Staccato_AN/app/src/main/java/com/on/staccato/data/network/TokenManager.kt
@@ -1,30 +1,35 @@
 package com.on.staccato.data.network
 
-import com.on.staccato.StaccatoApplication
+import com.on.staccato.data.member.MemberDataSource
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import javax.inject.Inject
 
-class TokenManager {
-    private var cachedToken: String? = null
+class TokenManager
+    @Inject
+    constructor(
+        private val memberLocalDataSource: MemberDataSource,
+    ) {
+        private var cachedToken: String? = null
 
-    suspend fun getToken(): String? {
-        return if (cachedToken.isNullOrEmpty()) {
-            fetchAndCacheToken()
-        } else {
-            cachedToken
+        suspend fun getToken(): String? {
+            return if (cachedToken.isNullOrEmpty()) {
+                fetchAndCacheToken()
+            } else {
+                cachedToken
+            }
+        }
+
+        private suspend fun fetchAndCacheToken(): String? {
+            val token =
+                withContext(Dispatchers.IO) {
+                    memberLocalDataSource.getToken()
+                }
+            cachedToken = token
+            return token
+        }
+
+        fun clearToken() {
+            cachedToken = null
         }
     }
-
-    private suspend fun fetchAndCacheToken(): String? {
-        val token =
-            withContext(Dispatchers.IO) {
-                StaccatoApplication.userInfoPrefsManager.getToken()
-            }
-        cachedToken = token
-        return token
-    }
-
-    fun clearToken() {
-        cachedToken = null
-    }
-}
