+++ android/Staccato_AN/app/src/main/java/com/on/staccato/domain/repository/MemberRepository.kt
@@ -9,5 +9,7 @@ interface MemberRepository {
 
     suspend fun getMemberId(): Result<Long>
 
+    suspend fun getNickname(): Result<String>
+
     suspend fun searchMembersBy(nickname: String): Flow<ApiResult<Members>>
 }
