+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/staccato/viewmodel/StaccatoViewModel.kt
@@ -4,12 +4,12 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.on.staccato.StaccatoApplication
 import com.on.staccato.data.network.onException
 import com.on.staccato.data.network.onServerError
 import com.on.staccato.data.network.onSuccess
 import com.on.staccato.domain.model.Feeling
 import com.on.staccato.domain.model.StaccatoShareLink
+import com.on.staccato.domain.repository.MemberRepository
 import com.on.staccato.domain.repository.StaccatoRepository
 import com.on.staccato.presentation.common.MutableSingleLiveData
 import com.on.staccato.presentation.common.SingleLiveData
@@ -18,8 +18,6 @@ import com.on.staccato.presentation.staccato.detail.StaccatoDetailUiModel
 import com.on.staccato.presentation.staccato.detail.StaccatoShareEvent
 import com.on.staccato.presentation.util.ExceptionState
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.Deferred
-import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -27,6 +25,7 @@ import javax.inject.Inject
 class StaccatoViewModel
     @Inject
     constructor(
+        private val memberRepository: MemberRepository,
         private val staccatoRepository: StaccatoRepository,
     ) : ViewModel() {
         private val _staccatoDetail = MutableLiveData<StaccatoDetailUiModel>()
@@ -48,24 +47,38 @@ class StaccatoViewModel
         val shareEvent: SingleLiveData<StaccatoShareEvent> get() = _shareEvent
 
         fun createStaccatoShareLink() {
-            val nickname = getUserNickname()
             val staccatoId = staccatoDetail.value?.id
             if (staccatoId == null) {
                 handleException(ExceptionState.UnknownError)
                 return
             }
+
             viewModelScope.launch {
-                staccatoRepository.createStaccatoShareLink(staccatoId)
-                    .onSuccess { postShareEvent(nickname.await(), it) }
-                    .onException(::handleException)
-                    .onServerError(::handleServerError)
+                val nickname: String = getUserNickname() ?: return@launch
+                shareStaccato(staccatoId, nickname)
             }
         }
 
-        private fun getUserNickname(): Deferred<String> =
-            viewModelScope.async {
-                StaccatoApplication.userInfoPrefsManager.getMemberProfile().nickname
+        private suspend fun getUserNickname(): String? {
+            val nickname = memberRepository.getNickname()
+            if (nickname.isFailure) {
+                handleException(ExceptionState.UnknownError)
+                return null
             }
+            return nickname.getOrNull()
+        }
+
+        private suspend fun shareStaccato(
+            staccatoId: Long,
+            nickName: String,
+        ) {
+            staccatoRepository.createStaccatoShareLink(staccatoId)
+                .onSuccess {
+                    postShareEvent(nickName, it)
+                }
+                .onException(::handleException)
+                .onServerError(::handleServerError)
+        }
 
         private fun postShareEvent(
             nickname: String,
